+++ src/main/java/com/babinska/PlannerForTutor/lessonReservation/dto/LessonReservationDto.java
@@ -3,8 +3,8 @@ package com.babinska.PlannerForTutor.lessonReservation.dto;
 import com.babinska.PlannerForTutor.lessonReservation.LessonType;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 
 @Builder
@@ -15,6 +15,6 @@ public record LessonReservationDto (Long id,
                              LocalTime endTime,
                              LocalDate reservationDate,
                              int durationInMinutes,
-                             double price) {
+                             BigDecimal price) {
 }
